Soporte a lectura de elementos estadisticos de BD
Pendiente
+ Implementar en consulta:
  - Nombre de la entidad
  - Localizacion geografica de entidad
+ Agregar soporte a piramide poblacional y envio del POCO respectivo
+ Agregar archivo SQL de generacion de BD al repositorio

+++ modelo/PoblacionPorRangoDeEdad.h
@@ -6,9 +6,9 @@ class PoblacionPorRangoDeEdad
 {
 public:
     QString nombre;
-    int numeroDeHombres;
-    int numeroDeMujeres;
-    int totalDePoblacion;
+    double numeroDeHombres;
+    double numeroDeMujeres;
+    double totalDePoblacion;
 };
 
 #endif // POBLACIONPORRANGODEEDAD_H
