+++ vista/IDelegadoObjetoBurbuja.h
@@ -13,7 +13,7 @@ public:
 
     virtual void asignarColorAElemento(QColor color) = 0;
 
-    virtual void asignarRadioAElemento(int radio) = 0;
+    virtual void asignarRadioAElemento(double radio) = 0;
 
     virtual void establecerVisible(bool visible) = 0;
